Move tests to separate files.
