+++ .github/workflows/main.yml
@@ -13,4 +13,4 @@ jobs:
     steps:
       - uses: actions/checkout@v2
       - run: pip install pytest
-      - run: python -m pytest *.py
+      - run: python -m pytest
