+++ Makefile
@@ -1,4 +1,2 @@
 test:
-	python3 node.py
-	python3 texi2rst.py
-	python3 texi2xml.py
+	pytest
